Renamed sil header file/source files to prefix with sil_.
