+++ firmware/quadruna/sil/src/manager/sil_board.c
@@ -1,4 +1,4 @@
-#include "board.h"
+#include "sil_board.h"
 #include <stdio.h>
 #include <czmq.h>
