+++ firmware/quadruna/sil/src/manager/sil_board.h
