+++ firmware/quadruna/sil/src/shared/sil_atoi.h
